dashboard/requirements.txt
dash==2.14.2
plotly==5.18.0
dash-bootstrap-components==1.5.0
influxdb-client==1.38.0
pandas==2.1.4
PyYAML==6.0.1
